fixed float cast
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/trunk@7571 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ registration/include/pcl/registration/impl/correspondence_estimation_organized_projection.hpp
@@ -103,7 +103,7 @@ pcl::registration::CorrespondenceEstimationOrganizedProjection<PointSource, Poin
 
         double dist = (p_src3 - (*target_) (u, v).getVector3fMap ()).norm ();
         if (dist < max_distance)
-          correspondences[c_index ++] =  pcl::Correspondence (*src_it, v * target_->width + u, dist);
+          correspondences[c_index ++] =  pcl::Correspondence (*src_it, v * target_->width + u, static_cast<float> (dist));
       }
     }
   }
